+++ .gitignore
@@ -129,7 +129,7 @@ dmypy.json
 # tests
 .tests/
 test_data/
-padeopsIO/test*
+padeopsio/test*
 test*
